# src/Web.Client/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/product-development": {
      "get": {
        "tags": [
          "ProductDevelopment"
        ],
        "summary": "Получение всех продуктов в разработке",
        "operationId": "GetAllProductDevelopments",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 1
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 2
          },
          {
            "name": "filterExpression",
            "in": "query",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 3
          },
          {
            "name": "filterValues",
            "in": "query",
            "style": "form",
            "explode": true,
            "schema": {
              "type": "array",
              "nullable": true,
              "items": {
                "type": "string",
                "nullable": true
              }
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDevelopmentResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ProductDevelopment"
        ],
        "summary": "Заполнение состава продукта",
        "operationId": "SetCompositionProduct",
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompositionProductRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/product-development/issue/{issueId}": {
      "get": {
        "tags": [
          "ProductDevelopment"
        ],
        "summary": "Получение всех продуктов по ID задачи",
        "operationId": "GetProductDevelopmentByIssueId",
        "parameters": [
          {
            "name": "issueId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          },
          {
            "name": "skip",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            },
            "x-position": 2
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 100
            },
            "x-position": 3
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductDevelopmentResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/product-development/{productDevelopmentId}/complete-issue": {
      "put": {
        "tags": [
          "ProductDevelopment"
        ],
        "summary": "Завершить задачу, переход на следующий этап или создание готового продукта",
        "operationId": "CompleteIssueByProductDevelopment",
        "parameters": [
          {
            "name": "productDevelopmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/product-development/{productDevelopmentId}/composition": {
      "get": {
        "tags": [
          "ProductDevelopment"
        ],
        "summary": "Получение состава продукта по ProductDevelopmentId",
        "operationId": "GetProductCompositionByProductDevelopmentId",
        "parameters": [
          {
            "name": "productDevelopmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompositionProductGroupModel"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/product-development/{productDevelopmentId}/techcard-items": {
      "get": {
        "tags": [
          "ProductDevelopment"
        ],
        "summary": "Получение состава тех.карты по ProductDevelopmentId",
        "operationId": "GetProductTechCardItemsByProductDevelopmentId",
        "parameters": [
          {
            "name": "productDevelopmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductDevelopmentResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "productId": {
            "type": "string",
            "format": "guid"
          },
          "productName": {
            "type": "string"
          },
          "manufacturingProcessId": {
            "type": "string",
            "format": "guid"
          },
          "batchNumber": {
            "type": "string"
          },
          "issueId": {
            "type": "string",
            "description": "ID задачи, где находится продукт",
            "format": "guid"
          },
          "operationName": {
            "type": "string"
          },
          "serialNumber": {
            "type": "string",
            "description": "Серийный номер продукта"
          },
          "settingOption": {
            "description": "Вариант настройки",
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/SettingOption"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ProductStatus"
          }
        }
      },
      "SettingOption": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "ЗВУ",
          "R",
          "ВДК",
          "Триада",
          "DC_DC"
        ],
        "enum": [
          100,
          200,
          300,
          400,
          500
        ]
      },
      "ProductStatus": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Initial",
          "InPaused",
          "InProgress",
          "Completed",
          "Cancelled"
        ],
        "enum": [
          100,
          101,
          200,
          300,
          400
        ]
      },
      "CompositionProductRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "productDevelopmentId": {
            "type": "string",
            "description": "Продукт относительно которого заполняем состав",
            "format": "guid"
          },
          "completedProductsId": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "guid"
            }
          }
        }
      },
      "CompositionProductGroupModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "productDevelopmentId": {
            "type": "string",
            "format": "guid"
          },
          "parentProductName": {
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          },
          "completedProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CompletedProductCompositionModel"
            }
          }
        }
      },
      "CompletedProductCompositionModel": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "completedProductId": {
            "type": "string",
            "format": "guid"
          },
          "itemName": {
            "type": "string"
          },
          "serialNumber": {
            "type": "string"
          }
        }
      },
...
    }
  }
}